fix: address cppcheck static analysis issues
- fixed constVariablePointer in copy ITCM/DTCM function
- reduced scope of index variable in add_co2_to_queue
- fixed printf format specifier for uint32_t
- suppressed constParameterCallback in co2_task

+++ Core/Src/stm32h7xx_it.c
@@ -319,6 +319,6 @@ void HardFault_Handler_C(uint32_t *stacked_regs) {
 
 static void print_reg(const char *name, uint32_t val)
 {
-    printf("%s = 0x%08lX\r\n", name, val);
+    printf("%s = 0x%08lX\r\n", name, (unsigned long)val);
 }
 /* USER CODE END 1 */
